+++ packages/atom.io/__tests__/public/families.test.ts
@@ -1,12 +1,5 @@
 import type { Logger } from "atom.io"
-import {
-	atomFamily,
-	getState,
-	makeRootMoleculeInStore,
-	selectorFamily,
-	setState,
-	Silo,
-} from "atom.io"
+import { atomFamily, getState, selectorFamily, setState } from "atom.io"
 import { findState } from "atom.io/ephemeral"
 import * as Internal from "atom.io/internal"
 import { vitest } from "vitest"
