+++ packages/atom.io/internal/src/index.ts
@@ -33,7 +33,7 @@ export * from "./junction"
 export * from "./keys"
 export * from "./lazy-map"
 export * from "./lineage"
-export type * from "./molecule"
+export * from "./molecule"
 export * from "./mutable"
 export * from "./not-found-error"
 export * from "./operation"
