+++ packages/atom.io/src/dispose-state.ts
@@ -3,13 +3,23 @@ import type { Canonical } from "atom.io/json"
 
 import type { ReadableFamilyToken, ReadableToken } from "."
 
+/**
+ * @public
+ * Disposes of a state in the implicit store
+ * @param token - The token of the state to dispose
+ * @overload Default
+ */
 export function disposeState(token: ReadableToken<any>): void
-
+/**
+ * @public
+ * Disposes of a state family in the implicit store
+ * @param token - The token of the state family to dispose
+ * @param key - The unique key of the state to dispose
+ */
 export function disposeState<K extends Canonical>(
 	token: ReadableFamilyToken<any, K>,
 	key: K,
 ): void
-
 export function disposeState(
 	...[token, key]:
 		| [token: ReadableFamilyToken<any, any>, key: Canonical]
